feat(deposit-components): move to general to reuse
