+++ src/pages/ValidatorManagement/ManageValidator/RightManageSettings/Panels/DepositPanel.tsx
@@ -6,8 +6,8 @@ import PanelWrapper from './PanelWrapper'
 import { RootState } from '../../../../../redux/store'
 import ConnectedWallet from '../../../../../components/General/ConnectedWallet'
 import ConnectWallet from '../../../../../components/General/ConnectWallet'
-import ChainParity from '../../../../ValidatorOnboarding/Deposit/ChainParity'
-import ValidatorRequest from '../../../../ValidatorOnboarding/Deposit/ValidatorRequest/ValidatorRequest'
+import ChainParity from '../../../../../components/General/ChainParity'
+import ValidatorRequest from '../../../../../components/General/ValidatorRequest/ValidatorRequest'
 import { getDepositTitle } from '../../../../../utilities'
 import DepositValidator from './DepositValidator'
 import ValidatorBalanceInfoBox from '../../../../ValidatorOnboarding/Deposit/ValidatorBalanceInfoBox'
